solved the issue of getting the accesstoken when creating a new user, to work on user profile bug
