+++ src/components/Navbar.jsx
@@ -1,30 +1,11 @@
-import React, { useEffect } from "react";
+import React from "react";
 import "./navbar.css";
 import { Link } from "react-router-dom";
 import { NavLogo, LessonPNG, CharacterPNG, ProfilePNG, VoicesPNG } from "./PNG";
 import { useAuth0 } from "@auth0/auth0-react";
-import axios from "axios";
-import { Backend_URL } from "../BACKEND_URL.js";
 
 export default function Navbar(userData) {
-  const { user, isAuthenticated } = useAuth0();
-
-  useEffect(() => {
-    if (isAuthenticated) {
-      const userInfo = {
-        username: user?.nickname,
-        first_name: user?.given_name || user?.name,
-        last_name: user?.family_name || "",
-        email_address: user?.email,
-        profile_pic_url: user?.picture,
-        voicevox_id: 1,
-      };
-      axios.post(`${Backend_URL}/users/newUser`, userInfo).catch((err) => {
-        console.log(userInfo);
-        console.log("Axios post to BE error", err);
-      });
-    }
-  }, [isAuthenticated]);
+  const { isAuthenticated } = useAuth0();
 
   return (
     <div className="navbar">
@@ -81,40 +62,3 @@ export default function Navbar(userData) {
     </div>
   );
 }
-
-//<<<<<<<<<<<<<<SETTLE NEXT TIME(METADATA)>>>>>>>>>>>>>>>>>>>
-// const [userMetadata, setUserMetadata] = useState({});
-// const getUserMetaData = async () => {
-//   const domain = process.env.REACT_APP_AUTH0_DOMAIN;
-//   // const audience = process.env.REACT_APP_AUTH0_AUDIENCE;
-//   try {
-//     const accessToken = await getAccessTokenSilently({
-//       // authorizationParams: {
-//       audience: `https://${domain}/api/v2/`,
-//       scope: "read:users",
-//       // },
-//     });
-//     const userDetailsByIdUrl = `https://${domain}/api/v2/users/${user.sub}`;
-//     const metadataResponse = await fetch(userDetailsByIdUrl, {
-//       headers: {
-//         Authorization: `Bearer ${accessToken}`,
-//       },
-//     });
-//     console.log("metadata response", metadataResponse);
-//     const { user_metadata } = await metadataResponse.json();
-//     console.log("user_metadata", user_metadata);
-//     setUserMetadata(user_metadata);
-//     console.log(
-//       "authenticated",
-//       isAuthenticated,
-//       "user",
-//       user,
-//       "metadata",
-//       userMetadata,
-//       "user.sub",
-//       user.sub
-//     );
-//   } catch (err) {
-//     console.log("error", err.message);
-//   }
-// };
